% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{append_design_matrix}
\alias{append_design_matrix}
\title{Prepare the data matrix by appending a column of 1's top it}
\usage{
append_design_matrix(df)
}
\arguments{
\item{df}{Data frame, minus the response}
}
\description{
Prepare the data matrix by appending a column of 1's top it
}
\examples{
append_design_matrix(matrix(c(1,2,3,4,5,6), nr=2, byrow=TRUE))

df<-iris
df<-df[,2:dim(df)[2]] # remove first column, using it as the response vector
append_design_matrix(df)

}
