% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{betah}
\alias{betah}
\title{Point estimate for beta, used in linear regression}
\usage{
betah(X, Y)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Response vector}
}
\description{
Point estimate for beta, used in linear regression
}
\examples{
betah(matrix(c(1,2,3,4,5,6), nr=2, byrow=TRUE),
      matrix(c(1,2,3), nc=3))

}
