% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{bootreg}
\alias{bootreg}
\title{Perform boot strapping}
\usage{
bootreg(Y, X, iter = 1, alpha = 0.05)
}
\arguments{
\item{Y}{response vector}

\item{X}{design matrix}

\item{alpha}{(=0.05) significance level for generating confidence intervals}

\item{iter(=1)}{number of iterations to run bootstrap}
}
\description{
Perform boot strapping
}
\examples{
df<-iris

# scrape off all non-numeric data first
df<-df[,sapply(df, is.numeric)]
df
X<-df[,-1]
Y<-as.matrix(df[,1])
X
Y
dim(X)
dim(Y)
X<-append_design_matrix(X)
dim(X)
bh=bootreg(Y,X,1000) # works just fine! Whoo!
bh
}
