% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterReduction.R
\name{cluster_k_analysis}
\alias{cluster_k_analysis}
\title{Perform the types of anaylsis which do a reduction on the data set,
finding the optimum value of k}
\usage{
cluster_k_analysis(x, kmax = 10, elbowPlotFileName = "elbowPlot.jpeg",
  silhouetteFileName = "silhouette.jpeg", gapFileName = "gap.jpeg",
  iter.max = 100, nstart = 1, algorithm = "Lloyd", trace = FALSE,
  debugOutput = FALSE, debugOutFileNameBase = "sil")
}
\arguments{
\item{x}{Data in the form of a data frame or data matrix, for safety, please use data.matrix(...) as a wrapper}

\item{kmax}{(=10) Maximum number of clusters to form}

\item{elbowPlotFileName}{(="elbowPlot.jpeg"), name of file for elbow plot output}

\item{silhouetteFileName}{(="silhouette.jpeg"), name of file for silhouette plot output}

\item{gapFileName}{(="gap.jpeg"), name of file for output of gap statistics plot}

\item{iter.max}{(=100) Maximum number of iterations}

\item{nstart}{How many random sets to be chosen}

\item{algorithm}{What algorithm to use (default="Lloyd").
choices include: "Hartigan-Wong", "LLoyd", "MacQueen", and "Forgy"}

\item{trace}{logical or integer number, currently only used in the default method ("Hartigan-Wong"):
if positive (or true), tracing information on the progress of the algorithm is produced.
Higher values may produce more tracing information.}

\item{debugOutput}{(=FALSE) Boolean value for whether or not to output silhouette plots for
each and every cluster value used. This generates a TON of output! YOU HAVE BEEN WARNED!}

\item{debugOutFileNameBase}{(="sil") Base file name for debug output, if enabled.}
}
\description{
NOTE: by default, this will perform the analysis on the scaled data
}
\examples{
cluster_k_analysis(data.matrix(iris), kmax=19)
cluster_k_analysis(data.matrix(iris), kmax=19, iter.max=200, debugOutput=T)
}
