% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{determine_greatest_contributors}
\alias{determine_greatest_contributors}
\title{Determine greatest data factors}
\usage{
determine_greatest_contributors(x, k)
}
\arguments{
\item{x}{Data set to be analyzes}

\item{k}{Number of components considered}
}
\description{
Determine greatest data factors
}
\examples{
determine_greatest_contributors(data.matrix(iris), 2)

}
