\name{ggfast-package}
\alias{ggfast-package}
\alias{ggfast}
\docType{package}
\title{
\packageTitle{ggfast}
}
\description{
\packageDescription{ggfast}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggfast}
\packageIndices{ggfast}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggfast}

Maintainer: \packageMaintainer{ggfast}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
