% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCAPlot}
\alias{PCAPlot}
\title{After calling whatever means of cluster reductions, now do a PCA
on the already reduced data set.}
\usage{
PCAPlot(x, k, barPlotFileName = "barPlotComp.jpeg")
}
\arguments{
\item{x}{Data set to be analyzed}

\item{k}{Number of components considerd}

\item{barPlotFileName}{(="barPlotComp.jpeg") file name used to store bar plot output}
}
\description{
After calling whatever means of cluster reductions, now do a PCA
on the already reduced data set.
}
\examples{
PCAPlot(data.matrix(iris), 2)

}
