% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{predictor}
\alias{predictor}
\title{Predict the next value using the same multilinear regression model}
\usage{
predictor(X, model)
}
\arguments{
\item{X}{design matrix}

\item{model}{the result of running bootreg on the data set, ie. a great big collection
of different betas, upon which the column mean is used to get the most accurate possible
estimate for beta}
}
\description{
Predict the next value using the same multilinear regression model
}
\examples{
# predict a singular value, given the description of the linear model and the design matrix of the new points
df<-iris

# scrape off all non-numeric data first
df<-df[,sapply(df, is.numeric)]
df
X<-df[,-1]
Y<-as.matrix(df[,1])
X
Y
dim(X)
dim(Y)
X<-append_design_matrix(X)
dim(X)
bh=bootreg(Y,X,1000) # works just fine! Whoo!
data.matrix(rowMeans(bh$bh))
predictor(matrix(c(1, 2, 2, 3), nr=1), bh)

}
