% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterReduction.R
\name{reduce_cluster}
\alias{reduce_cluster}
\title{Reduce the amount of data through using kmeans clustering algorithm
Call the in R kmean cluster function in order to reduce a pool of data into a few, clustered points}
\usage{
reduce_cluster(x, k, fileName, iter.max = 10, nstart = 1,
  algorithm = "Lloyd", trace = FALSE)
}
\arguments{
\item{x}{Data in the form of a data frame or data matrix, for safety, please use data.matrix(...) as a wrapper}

\item{k}{Number of clusters to form}

\item{fileName}{Name of file you wish to save a .csv formatted file to}

\item{iter.max}{Number of iterations to use at a maximum (default = 10)}

\item{nstart}{How many random sets to be chosen}

\item{algorithm}{What algorithm to use (default="Lloyd").
choices include: "Hartigan-Wong", "LLoyd", "MacQueen", and "Forgy"}

\item{trace}{logical or integer number, currently only used in the default method ("Hartigan-Wong"):
 if positive (or true), tracing information on the progress of the algorithm is produced.
 Higher values may produce more tracing information.

 @examples
 reduce_cluster(data.matrix(iris), 10, "myFile.csv")
 reduce_cluster(df, 15, "myManyIterationsFile.csv", iter.max=1000, algorithm="MacQueen")}
}
\description{
Reduce the amount of data through using kmeans clustering algorithm
Call the in R kmean cluster function in order to reduce a pool of data into a few, clustered points
}
