% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterReduction.R
\name{visualize_reduction}
\alias{visualize_reduction}
\title{Visualize the amount of data reduction in the first component from using clustering
This verifies that the clustering analysis can be of some use for data reduction}
\usage{
visualize_reduction(x, k)
}
\arguments{
\item{x}{Data in the form of a data frame or data matrix, for safety, please use data.matrix(...) as a wrapper}

\item{k}{Number of clusters to form}
}
\description{
Visualize the amount of data reduction in the first component from using clustering
This verifies that the clustering analysis can be of some use for data reduction
}
\examples{
visualize_reduction(data.matrix(iris), 19)
}
